Renamed openapi files.
